project-root/backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.0
gunicorn>=20.1
psycopg2-binary>=2.9 ; python_version >= "3.8"  # optional, only if you switch to Postgres
